Lazily rebuild timelinePropertyCache only when needed
This commit avoids an apparent O(n^2) complexity problem when adding
keyframes.  Each keyframe results in the recomputation of the property
cache, which takes an increasing amount of time with the addition of
each keyframe.  Delaying this recomputation until it is required (in
_updateState) changes the page load time for a few hundred keyframes
from around 10 seconds to nearly instant for me.