New _timelinePropertyCache structure
This changes _timelinePropertyCache to be a sorted array, and removes
the _timelinePropertyCacheKeys as that is unnecessary since the cache
entries are directly sorted.

This also uses a binary search to find cache entries rather than a
linear search, and has a new algorithm for creating the cache that
is faster.

All in all these changes make generating the cache about 4x faster,
and should also be faster at accessing the cache due to a simpler
structure and the aforementioned binary search.