Warnings/Errors for common keyframe screwups
Duplicate keyframes always cause an error as they break playback.  Set
`rekapi._warnOnOutOfOrderKeyframes` to warn on out-of-order keyframe
assignments (i.e. before the last one on the track) as this can indicate
a keyframe overlap that can cause strange animation effects if not
planned for.  `new Error` is used even in the warning case to provide
a backtrace.