Correct gPCEFMillisecond to always work for times < first keyframe
I had wanted to restructure this as such when I did the comment/return
fix, but incorrectly decided on the other path to reduce potential
breakage.  The only reason it was working before was that `-1` was
something that you could pull properties out of, even though it is
distinctly not a cache entry.

This will still cause an error if the timeline is empty.